move the sampler to the base module
