+++ doc/miscellaneous.rst
@@ -4,7 +4,7 @@
 Miscellaneous samplers
 ======================
 
-.. currentmodule:: imblearn.misc
+.. currentmodule:: imblearn
 
 .. _function_sampler:
 
@@ -18,7 +18,7 @@ a single function. Additional parameters can be added using the attribute
 to retain the 10 first elements of the array ``X`` and ``y``::
 
   >>> import numpy as np
-  >>> from imblearn.misc import FunctionSampler
+  >>> from imblearn import FunctionSampler
   >>> from sklearn.datasets import make_classification
   >>> X, y = make_classification(n_samples=5000, n_features=2, n_informative=2,
   ...                            n_redundant=0, n_repeated=0, n_classes=3,
